Initial work towards modularised deps
